restructured package to incorporate __init__.py file
