+++ flask_test/app.py
@@ -8,50 +8,6 @@ app.config['SECRET_KEY'] = '789531f82597c4e6f638a3c40ac94a6c'
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
 db = SQLAlchemy(app)
 
-class Business(db.Model):
-    """Create class 'Business' for Yelp App
-    Business has methods:
-    - id
-    - name
-    - location
-    - review1
-    - review1_id
-    - review2
-    - review2_id
-    - review3
-    - review3_id
-    
-    added __repr__ function to show content of Business class as text (vs. as location in memory)"""
-    
-    id = db.Column(db.String, primary_key = True)
-    name = db.Column(db.String())
-    location = db.Column(db.String())
-    review1 = db.Column(db.String())
-    review1_id = db.Column(db.String())
-    review2 = db.Column(db.String())
-    review2_id = db.Column(db.String())
-    review3 = db.Column(db.String())
-    review3_id = db.Column(db.String())
-    
-    def __repr__(self):
-        return '<NAME {}>'.format(self.name)
-
-
-# create review class for sqlite database
-class Review(db.Model):
-    """Create class 'Review' for Yelp App
-    Business has methods:
-    - id
-    - text
-    
-    added __repr__ function to show content of Review class as text (vs. as location in memory)"""
-    
-    id = db.Column(db.String, primary_key = True)
-    text = db.Column(db.String())
-    
-    
-
-
 # dummy review data for flask app display testing
 reviews = [
     
@@ -119,23 +75,3 @@ def businesses():
 # allow app to run using "python app.py" in terminal
 if __name__ == '__main__':
     app.run(debug=True)
-
-# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db.sqlite3'
-# app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-
-# DB = SQLAlchemy(app)
-
-# class Business(DB.Model):
-#     """Create class 'Business' for Yelp App
-#     Business has methods:
-#     - id
-#     - name
-#     - location
-    
-#     added __repr__ function to show content of Business class as text (vs. as location in memory)"""
-    
-#     id = DB.Column(DB.String, 
-#                    primary_key = True)
-#     name = DB.Column(DB.String(100), 
-#                      nullable = False)
-#     business_location = DB.Column(DB.String(100))
