+++ flask_test/flask_test/templates/register.html
